requirements.txt
# 依赖库
allure-pytest=2.14.0
jsonpath=0.82.2
pytest-rerunfailures=14.0
PyMySQL>=1.1.1
PyYAML>=6.0.1
requests~=2.32.3
pytest>=6.0.1